Flexbox, and horizontal layout on wide screens.
Where available, we use flexbox for layout as it produces a better
result, especially on narrow viewports. This lets us do a nice
horizontal layout for screens that are sufficiently wide, too.

I’ve also dropped back on things like rem units for the position:
absolute; layout, because most of the browsers that are using it don’t
like rems.