Apply the editor theme to the results pane.
This is not quite perfect (most notably, Ambiance lacks the grain—I tried
implementing that, but getting it perfect is fiendishly difficult
cross-browser), but it is a fair sight better than nothing.

I’ve implemented rehighlighting of the pygments-highlighted code to fit
ACE’s class names, because ACE does an abysmal job of highlighting ASM
and doesn’t know LLVM IR. So now the ASM/LLVM IR matches the theme too!

While I was at it I added a × button to the top right of the results
pane to close it.