Improve compiler output with things like colours.
The compiler can write ANSI colour codes for nicer error messages, so
why not use it? I improved a few other cases (like internal server
errors which I unsurprisingly triggered while developing—reporting such
as connection failures was wrong) while I was at it.

The colour part of it is mildly fragile, as what codes actually get
written depends on the terminfo. Most notably, the `^[(B` that seems to
come after `^[[m` (turn off all attributes) has baffled me as I haven’t
found documentation of what it is, nor do I find it in my termcap for
sgr0. Ah well. It doesn’t seem to matter. Hopefully the deployment
server will produce the same ANSI codes. If it doesn’t, the ansi2html
function may need to be adjusted.

Error codes are hyperlinks to their entries in the error index, which is
generated from the same stuff that the hitherto advised `rustc --explain
E0000` is.