Set TERM=xterm to get colours working well.
This should make colour codes actually work on the server (where `TERM`
is presumably unset, based on experience) and ensures stability for the
colour codes emitted, constraining it to the basic 16 colour palette
which is easier to deal with and more familiar to people (`^[[31m` for
red, &c.).

That unknown `^[(B` still appears, though now it’s *before* the `^[[m`,
unlike after as it was with my own `TERM=rxvt-unicode-256color`.
*Shrug.*