Support standard `ansi2html` escape delimiter
I did a bit of research, and `[0m` seems to be the standard "reset format" ANSI token. I can't find much information on using a simple `[m` instead, but it must is supported by many terminals.

Anyway, rustc switched from `[m` to `[0m` in a recent nightly. This broke `ansi2html`'s regex, which was only expecting `[m`.

This should address and close #157.